Now handles ROMS grids using Cartesian coordinates
Modified to recgonize the difference between a "spherical" ROMS grid
(one which stores coordaintes as lat/lon) and a non-spherical ROMS
grid, which stores coordinates in x/y).  Computes dx, dy, angle, and
area differently depending on which type of coordinates are used.

(Still remaining: generating the "quick mosaic" files.)