Now writes coupler exchange grid files
Added ability to output the <type1>_mosaic_tile1X<type2>_mosaic_tile1.nc
exchange grid files, where <type1> and <type2> are various
combinations of "atmos", "land", and "ocean".  For testing, in order
to match the output of the original make_quick_mosaic tool, I had to
match the computation of cell area done by that tool.  This is
temporary, and I'll soon make the cell areas more consistent with the
areas from the supergrid.